Move safe_content to utilities
