+++ app/utilities/__init__.py
@@ -1,0 +1,1 @@
+from .strings import safe_content
